refactor(swap): simplify swap state organization
