+++ apps/mobile/src/features/swap/swap-state/strategies/execution-type/execution-type.utils.ts
